fix: public svg moved
