+++ my-monorepo/apps/shopease/src/pages/login/index.tsx
@@ -1,5 +1,5 @@
 import { useLogin } from "@/features/auth/hooks/useLogin";
-import google from "../../../public/google.svg";
+import google from "../../app/assets/google.svg";
 import { useNavigation } from "@/shared/hooks/useNavigation";
 import { z } from "zod";
 import { zodResolver } from "@hookform/resolvers/zod";
